refactor: split training/__init__.py into per-trainer modules
Move 1,950-line monolithic __init__.py into focused modules:
- _base.py: TrainConfig, Trainer ABC, shared CUDA/subprocess utilities
- groot.py: Gr00tTrainer
- lerobot.py: LerobotTrainer
- dreamgen.py: DreamgenIdmTrainer, DreamgenVlaTrainer
- cosmos_predict.py: CosmosTrainer
- cosmos_transfer.py: CosmosTransferTrainer
- evaluate.py: standalone evaluation harness
- __init__.py: re-exports and create_trainer() factory

Deduplicate _build_subprocess_env (was copy-pasted in both Cosmos
trainers) into shared _build_cosmos_subprocess_env in _base.py.

Public API unchanged - all existing imports continue to work.
